Move utils to _internal
